+++ Games/BrickBreakerPlus/main.cpp
@@ -299,7 +299,7 @@ public:
                 const int y = (BRICK_TOP + r * 2) * tileSize;
                 SDL_Rect rect = {x + sx + 1, y + sy + 1,
                                  tileSize * BRICK_W - 2, tileSize - 2};
-                SDL_SetRenderDrawColor(sdl, 255, 90 + r * 30, 90 + r * 20, 255);
+                SDL_SetRenderDrawColor(sdl, 255, (Uint8)(90 + r * 30), (Uint8)(90 + r * 20), 255);
                 SDL_RenderFillRect(sdl, &rect);
             }
         }
